backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4

# Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Supabase (acesso server-side)
supabase==2.10.0

# Variáveis de ambiente
python-dotenv==1.0.1